NNS1-3092: Sorting utils for neurons table (#4968)
# Motivation

We'll want to provide custom sorting in the neurons table by clicking on
the column headers.
So we'll need to be able to sort based on different criteria.
In this PR we provide some utilities for custom sorting.

# Changes

1. Implement neuron comparators based on stake, dissolve delay and
neuron ID. We'll add more later.
2. Implement a neuron sorting function which takes an ordered list of
comparators to define the sorting order.

# Tests

Unit tests are added.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet