NNS1-3092: Sort neurons in table (#4971)
# Motivation

Neuron cards are sorted by stake and then dissolve delay.
We want the same for the neurons table until we implement custom
sorting.

Currently we tie break on creation time but:
1. That's not guaranteed to be unique because on the IC time is constant
for the duration of a message
2. Is not visible in the UI

So we decided to tie break on neuron ID instead.

# Changes

1. Order neurons by decreasing stake, then decreasing dissolve delay,
then increasing ID.

# Tests

1. Unit tests added and updated.
2. Manually at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not yet