NNS1-3093: Convert SNS neurons for table rendering (#4977)
# Motivation

We want to use the neurons table for SNS neurons as well.
The neurons table takes neurons in a specific `TableNeuron` type.
So we need to convert `SnsNeuron`s to `TableNeuron`s.

# Changes

Add utility function `tableNeuronsFromSnsNeurons` to convert
`SnsNeuron`n to `TableNeuron`s.

# Tests

1. Conveniently set the neuron state in `createMockSnsNeuron` if it
wasn't provided but a dissolve delay was specified.
2. Add unit tests for `tableNeuronsFromSnsNeurons`.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet