NNS1-3092: Add state field to TableNeuron (#4986)
# Motivation

We want to render neuron state in the neurons table.
For this, we need the neuron state in the `TableNeuron` type which is
used to render the neurons table.

# Changes

1. Add `state` field to `TableNeuron` type.
2. Populate the state field in
`frontend/src/lib/utils/neurons-table.utils.ts` for both NNS and SNS
neurons.
3. Destructure `mockTableNeuron` in neurons in `NeuronsTable.spec.ts` so
new fields don't need to be added immediately in the test and can be
updated in a separate PR when we update the UI.
4. Remove `rowHref` from `mockTableNeuron` as it's an optional field and
it should be possible to destructure `mockTableNeuron` and not specify
`rowHref`.

# Tests

1. Unit tests added for `frontend/src/lib/utils/neurons-table.utils.ts`.
2. Tested manually in another branch which also has UI changes.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet