NNS1-3092: Add horizontal scrollbar to neurons table (#5002)
# Motivation

The neurons can be cut off on a narrow viewport and we plan to add even
more columns.
We need to make sure that the whole table can be reached.

# Changes

Change `overflow: hidden` to `overflow-y: auto` on the responsive table.

# Tests

Manually.
I tried to make a screen recording but somehow while recording the
scrollbar doesn't appear.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet