NNS1-3092: Use empty columns to space the neurons table (#5028)
# Motivation

Currently, the columns in the neurons table are spaced out somewhat
evenly by dividing targeting equal width per column.
But because the actual content in each column is not the same, this
means that the extra space each column get is not the same.
If we add extra empty columns, we can make each column take up only as
much space as it needs, and divide the additional space evenly over the
empty columns.
This has the extra benefit that we can left-align a column without it
affecting where the additional space goes.

We have to make sure they don't add any extra space on the mobile
layout.

# Changes

1. Make the `cellComponent` of the responsive table column optional.
2. Change `--mobile-grid-template-areas` to only contain areas for
columns with a `cellComponent`.
3. Render cells for columns without `cellComponent` without content and
give them `display: none` on mobile viewport.
4. Add 3 extra empty rows in the neurons table. Give them template `1fr`
and change the template of the content columns from `minmax(max-content,
1fr)` to `max-content`.

# Tests

1. Unit test updates.
2. Manually tested at
https://qsgjb-riaaa-aaaaa-aaaga-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=qsgjb-riaaa-aaaaa-aaaga-cai

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary