Change dfx state directory logic (#5044)
# Motivation

When extracting an snsdemo snapshot, we put the data in a directory in
the dfx cache.
There are several problems with this:
1. The snapshot determines which version of dfx it is run with. But we
don't know this until after the state is extracted. So when we determine
the dfx cache directory, it might be a cache directory for a different
version than the snapshot is created for.
2. This also means that if your dfx version is updated, the install
script will look for the state in a different cache directory and not
find it. This means you get a clean snapshot while you may have wanted
to reuse the state you've created earlier.
3. On DevEnv, the cache directory exists on a different disk volume.
This means that moving the directories in place is as slow as copying.

# Changes

1. Search across all dfx cache directories to find the dfx state
directory, preferring newer versions.
2. If no existing state is found, or if it's deleted because of
`--clean`, put the state directory in `$HOME/.snsdemo-state` (regardless
of dfx version) instead of in the dfx cache.

# Tests

Manually with existing state directories in different places and without
existing state directory.

# Todos

- [x] Add entry to changelog (if necessary).