Sort all Svelte imports (#5089)
# Motivation

Make it easier to sort future imports without causing large diffs.

Note: Keeping the imports sorted is not enforced. We can consider that
separately if we want.

Tip: Don't review the changes in this PR. If you want to verify it, just
check out `main` and run `scripts/sort-svelte-imports`, then compare
your branch with this branch.

# Changes

Ran `scripts/sort-svelte-imports`.

# Tests

CI passes.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary