Merge logic for first and middle column headers (#5098)
# Motivation

The rendering logic for the first column header in the `ResponsiveTable`
component is now identical to the logic for the middle columns. So we
can simply this by merging the logic.

# Changes

Remove the logic for the first column header and rely on the logic for
the middle columns.

# Tests

Existing tests pass.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary