Change neuron ID column title to 'Neurons' (#5105)
# Motivation

In mobile view, only the first column header of the neurons and tokens
tables is visible.
In this case "Neurons" fits better as table header than "Neuron ID".
So we agreed with UX to change the first column title of the neurons
table from "Neuron ID" to "Neurons" for both mobile and desktop view.

# Changes

Change the title of the neuron ID column to `"Neurons"`.

# Tests

1. Unit test updated.
2. Screenshots updated.

# Todos

- [x] Add entry to changelog (if necessary).