NNS1-3135: Move sorting from NeuronsTable to ResponsiveTable (#5104)
# Motivation

We want to be able to change the neurons table sorting by clicking on
the column headers.
This should be implemented in the generic `ResponsiveTable` component.
Currently the sorting order is determined in the `NeuronsTable`
component.
In this PR we move the sorting logic to the `ResponsiveTable` component.
This is a pure refactoring that will make it easier to implement the
custom sorting later.

# Changes

1. Add `id` and `comparator` fields to the `ResponsiveTableColumn` type.
The order will be specified via the column ID. This is in anticipation
of being able to store the order in local/remove storage, which is
easier with strings than with comparator functions.
2. Change the `sortTableData` function to take the `order` and `column`
params to determine the sorting order.
3. Add an optional `order` prop to the `ResponsiveTable` component to
specify the columns to sort by.
4. Add `comparator`s and `id`s for the neurons table columns that are
currently used for sorting.
5. Pass the desired `order` from the `NeuronsTable` to the `Responsive`
table.

# Tests

1. Unit tests added.
2. Tested manually, that nothing changes, also in the tokens table.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary