NNS1-3136: Add modal for mobile table sorting (#5137)
# Motivation

When sorting the table on desktop, you can click on the column headers.
For mobile we need a different UI with a modal to select a sorting
option.
This PR just adds that modal but does not connect it to the table yet.

Here is a demo of it integrated in the neurons table:
https://mecbw-6maaa-aaaaa-qabkq-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=mecbw-6maaa-aaaaa-qabkq-cai

<img width="565" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/ff5cad62-d09b-4e6c-931b-6d27447b600b">


# Changes

Add `frontend/src/lib/modals/common/ResponsiveTableSortModal.svelte`
component.
The modal binds and updates the `order` prop and closes as soon as a
selection is made.

# Tests

1. Unit test added.
2. Manually

# Todos

- [ ] Add entry to changelog (if necessary).
not yet