NNS1-3136: Add sorting modal in responsive table (#5142)
# Motivation

When sorting the table on desktop, you can click on the column headers.
For mobile we have a different UI with a modal to select a sorting
option added in https://github.com/dfinity/nns-dapp/pull/5137

This PR hooks up that modal in the table.

# Changes

1. Add an icon button in the upper right of the responsive table, if any
of the columns are sortable.
2. Open the sorting modal if the button is clicked.
3. Move the `mobile-only` and `desktop-only` styles so they can be used
to hide the sorting button on desktop.

# Tests

1. Unit tests added
2. Manually, also by UX, at
https://mecbw-6maaa-aaaaa-qabkq-cai.dskloet-ingress.devenv.dfinity.network/neurons/?u=mecbw-6maaa-aaaaa-qabkq-cai

# Todos

- [x] Add entry to changelog (if necessary).