NNS1-3094: Add staking path (#5148)
# Motivation

Currently the neurons view has a universe picker on the left.
We plan to make it more like the tokens table where there is no left-nav
but rather a first step where you choose the project and then a view
with a back button just for that project.

The table with projects with be at `/staking` and the neurons table ,
but without universe picker left-nav, will continue to be at `/neurons.

In this PR we just add the scaffolding for the `/staking` path.

# Changes

1. Add `AppPath.Staking` constant.
2. Add mostly empty route files for `/staking`.
3. Making the "Neurons Staking" tab navigate to `/staking` instead of
`/neurons?u=...` when `ENABLE_PROJECTS_TABLE` is enabled.

# Tests

1. Unit test added.
2. Tested manually.

# Todos

- [ ] Add entry to changelog (if necessary).
not yet