Reduce fallback neuron refreshing (#5180)
# Motivation

Staking or topping up an SNS neuron is a 2-step process:
1. Transfer tokens to the neuron account
2. Claim or refresh the neuron
If the process is interrupted after step 1, your tokens might go
missing. So we want to be able to recover from this.

Currently we check the neuron account balance of all your neurons every
time neurons are synced, for example when you navigate to the neurons
table.

We already added checking a neuron when you navigate to its details page
in https://github.com/dfinity/nns-dapp/pull/5169

So as long as we are able to recover newly staked neurons we can stop
doing this every time we sync neurons.

# Changes

1. Call `claimNextNeuronIfNeeded` from `SnsNeurons`.
2. Remove neuron checking code from `syncSnsNeurons`.

# Tests

1. Unit test added for `SnsNeurons.svelte`.
2. Unit test removed for `syncSnsNeurons`.
3. Tested manually by transferring tokens to a neuron account of a
neuron that doesn't exist yet and seeing it appear in the table.
4. Add default permissions on the `mockSnsNeuron` as this is what a
default non-hotkey neuron would have.

Drive-by: Explicitly set empty permissions on an ineligible neuron in
`SnsVotingCard.spec.ts` to make it clear that's why it's ineligible.

# Todos

- [x] Add entry to changelog (if necessary).