Store SNS neurons after querying for actionable proposals (#5182)
# Motivation

We load neurons for all SNSes for actionable proposals. If we put them
in the store they can be used to display on the neurons table etc.
There was a comment saying we shouldn't do this because it prevents
`syncSnsNeurons` from checking neurons but
1. It doesn't actually prevent anything, and
2. We no longer check neurons in `syncSnsNeurons`.

# Changes

1. Rename `loadNeurons` to `loadSnsNeurons` so the context is clearer
when the function is used in another file.
2. Use `loadSnsNeurons` to load neurons into the store instead of just
querying them in `actionable-sns-proposals.services.ts`.
3. Drive-by: Remove a comment from `syncSnsNeurons` about checking
neurons which it no longer does.

# Tests

Unit test adapted to check that the `snsNeuronsStore` is populated.

# Todos

- [ ] Add entry to changelog (if necessary).
not necessary